pyproject.toml
[project]
name = "gaussiancar"
version = "1.0"
description = "GaussianCaR: Gaussian Splatting for Efficient Camera-Radar Fusion"
authors = [
    { name = "Santiago Montiel-Marin" }
]
requires-python = ">=3.10,<3.13"
dependencies = [
    "addict>=2.4.0",
    "diff-gaussian-rasterization>=1.0",
    "efficientnet-pytorch>=0.7.1",
    "einops>=0.8.1",
    "flash-attn>=2.8.3",
    "fvcore>=0.1.5.post20221221",
    "gradio>=5.44.1",
    "hydra-core>=1.3.2",
    "imgaug>=0.4.0",
    "jaxtyping>=0.3.2",
    "lightning>=2.5.0",
    "matplotlib>=3.10.6",
    "moviepy>=2.2.1",
    "numpy<2.0.0",
    "nuscenes-devkit>=1.2.0",
    "peft>=0.18.0",
    "pillow>=11.3.0",
    "rootutils>=1.0.7",
    "segmentation-models-pytorch>=0.5.0",
    "setuptools>=77",
    "spconv-cu126>=2.3.8",
    "timm>=1.0.21",
    "torch==2.8.0",
    "torch-scatter==2.1.2+pt28cu126",
    "torchvision>=0.23.0",
    "transformers>=4.56.0",
    "wandb>=0.21.3",
]

[build-system]
requires = ["setuptools >= 77", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["gaussiancar"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126" },
]
torchvision = [
  { index = "pytorch-cu126" },
]
diff_gaussian_rasterization = [
  { path = "./gaussiancar/ops/diff-gaussian-rasterization" },
]

[tool.uv.extra-build-dependencies]
flash-attn = [{ requirement = "torch", match-runtime = true }]


[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true


[tool.uv]
find-links = [
  "https://data.pyg.org/whl/torch-2.8.0+cu126.html"
]
